+++ src/main/java/com/example/quaterback/api/feature/dashboard/dto/response/HourlyDischargeResponse.java
@@ -1,5 +1,5 @@
 package com.example.quaterback.api.feature.dashboard.dto.response;
 
-public record HourlyDischargeResponse(int hour, int dischargeKwh) {
+public record HourlyDischargeResponse(int hour, double dischargeKwh) {
 
 }
